+++ tests/unit/Badoo/fixtures/expected/php5.php
@@ -61,7 +61,14 @@ function replaceSomething($string){
 
 class SomeClass{
     
+    const VALUE = 1;
+    
     public $a = 1;
+    
+    public static function getValue(){if (false !== ($__softmocksvariableforcode = \Badoo\SoftMocks::isMocked(SomeClass::class, static::class, __FUNCTION__))) {$mm_func_args = func_get_args();$params = array();return eval($__softmocksvariableforcode);}/** @codeCoverageIgnore */
+        
+        return \Badoo\SoftMocks::getClassConst(self::class, 'VALUE');}
+    
     
     public function method($string){if (false !== ($__softmocksvariableforcode = \Badoo\SoftMocks::isMocked(SomeClass::class, static::class, __FUNCTION__))) {$mm_func_args = func_get_args();$params = array($string);return eval($__softmocksvariableforcode);}/** @codeCoverageIgnore */