+++ tests/unit/Badoo/SoftMocksTest.php
@@ -380,4 +380,22 @@ class SoftMocksTest extends \PHPUnit\Framework\TestCase
         //file_put_contents(__DIR__ . '/fixtures/expected/' . $filename, file_get_contents($result));
         $this->assertEquals(file_get_contents(__DIR__ . '/fixtures/expected/' . $filename), file_get_contents($result));
     }
+
+    /**
+     * @expectedException \RuntimeException
+     * @expectedExceptionMessage You will never see this message
+     */
+    public function testNotOk()
+    {
+        throw new \RuntimeException("You will never see this message");
+
+        $Mock = $this->getMock(MyTest::class, ['stubFunction']);
+        $Mock->expects($this->any())->method('stubFunction')->willReturn(
+            function () {
+                yield 10;
+            }
+        );
+    }
+
+    public function stubFunction() {}
 }