+++ CHANGELOG.md
@@ -6,6 +6,7 @@ There are next changes:
 
 - PHP 7.1 support (mostly nullable and void return type declarations);
 - update nikic/php-parser to 3.0.6;
+- fix bug with throwing from generators;
 - added tests for constants redefine.
 
 ## v1.2.0